# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShopNest API Documentation",
    "version": "1.0.0",
    "description": "API endpoints and specs for the ShopNest full-stack e-commerce system"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/products": {
      "get": {
        "summary": "Fetch all products (with search, filter, sorting, pagination)",
        "tags": ["Product"],
        "parameters": [
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Search in name and description" },
          { "name": "category", "in": "query", "schema": { "type": "string" }, "description": "Category slug or ID" },
          { "name": "minPrice", "in": "query", "schema": { "type": "number" } },
          { "name": "maxPrice", "in": "query", "schema": { "type": "number" } },
          { "name": "sortBy", "in": "query", "schema": { "type": "string" }, "description": "price, createdAt, name" },
          { "name": "sortOrder", "in": "query", "schema": { "type": "string" }, "description": "asc, desc" },
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Products retrieved successfully" }
        }
      },
      "post": {
        "summary": "Create a new product (Admin)",
        "tags": ["Product"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "slug", "price", "stock", "sku", "categoryId"],
                "properties": {
                  "name": { "type": "string" },
                  "slug": { "type": "string" },
                  "price": { "type": "number" },
                  "discountPrice": { "type": "number", "nullable": true },
                  "stock": { "type": "integer" },
                  "sku": { "type": "string" },
                  "image": { "type": "string" },
                  "categoryId": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Product created successfully" }
        }
      }
    },
...
  }
}